backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.6.2
click==8.1.7
colorama==0.4.6
deprecation==2.1.0
Flask==3.0.3
flask-cors
gotrue==2.4.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
postgrest==0.16.8
pydantic==2.7.4
pydantic_core==2.18.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==1.0.5
six==1.16.0
sniffio==1.3.1
storage3==0.7.6
StrEnum==0.4.15
supabase==2.5.1
supafunc==0.4.6
typing_extensions==4.12.2
websockets==12.0
Werkzeug==3.0.3
groq
pytest
pytest-mock
